Rename Mongo::Obfuscator to NosqlObfucsator
